OutlookMobileGCCRestrictionsEnabled Parameter
I added the following note as requested by CELA:
Note: If you are a GCC customer, by turning off OutlookMobileGCCRestrictionsEnabled, you are agreeing to Outlook Mobile connecting to features which are not part of Microsoft’s GCC Offering and are not GCC Compliant.
Let me know if you have any feedback about that particular item. We just don't want to be legally responsible if their admin chooses to expose their tenant to services outside of the GCC boundary.